% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_barras.R
\name{grafico_barras}
\alias{grafico_barras}
\title{Gráfico de barras comparativo región vs país}
\usage{
grafico_barras(
  df_region,
  df_pais,
  comuna,
  cant,
  valor,
  ejey = "Porcentaje de cobertura (N° de XXX)",
  titulo = "Porcentaje de cobertura O'Higgins - XXX",
  comunas_requeridas = c("Chépica", "Chimbarongo", "Codegua", "Coinco", "Coltauco",
    "Doñihue", "Graneros", "La Estrella", "Las Cabras", "Litueche", "Lolol", "Machalí",
    "Malloa", "Marchihue", "Mostazal", "Nancagua", "Navidad", "Olivar", "Palmilla",
    "Paredones", "Peralillo", "Peumo", "Pichidegua", "Pichilemu", "Placilla", "Pumanque",
    "Quinta de Tilcoco", "Rancagua", "Rengo", "Requínoa", "San Fernando", "San Vicente",
    "Santa Cruz", "Región")
)
}
\arguments{
\item{df_region}{Dataframe con los valores de la región}

\item{df_pais}{Dataframe con los valores del país}

\item{comuna}{Nombre de la columna con las comunas}

\item{cant}{Nombre de la columna con los valores absolutos}

\item{valor}{Nombre de la columna con los valores proporcionales (0-1)}

\item{ejey}{Título del eje Y}

\item{titulo}{Título del gráfico}

\item{comunas_requeridas}{Vector con los nombres de comunas a validar}
}
\value{
Objeto plotly
}
\description{
Esta función genera un gráfico interactivo comparando comunas de la región y la cobertura del país.
}
