% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_estadistico.R
\name{mapa_estadistico}
\alias{mapa_estadistico}
\title{Genera un mapa estadístico interactivo para las comunas}
\usage{
mapa_estadistico(
  df_mapa,
  nombre_comuna,
  valor,
  titulo = "Mapa estadístico",
  tipo = c("continuo", "discreto"),
  colores = c("red1", "yellow1", "green3"),
  titulo_leyenda = "Valor",
  comunas_requeridas = NULL,
  tooltip_text = NULL
)
}
\arguments{
\item{df_mapa}{Data frame con la información a mapear.}

\item{nombre_comuna}{Nombre de la columna con el nombre de la comuna (string).}

\item{valor}{Nombre de la columna con el valor numérico a representar (string).}

\item{titulo}{Título del mapa.}

\item{tipo}{Tipo de escala: "continuo" o "discreto".}

\item{colores}{Vector de colores para la escala.}

\item{titulo_leyenda}{Título de la leyenda.}

\item{comunas_requeridas}{Vector opcional para forzar comunas específicas.}

\item{tooltip_text}{Texto plantilla opcional para personalizar el tooltip, ej: "Comuna: {nombre_comuna}\if{html}{\out{<br>}}N° casos: {n}\if{html}{\out{<br>}}Tasa: {valor}"}
}
\value{
Un objeto plotly.
}
\description{
Genera un mapa estadístico interactivo para las comunas
}
