% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_tabla.R
\name{rt_tabla}
\alias{rt_tabla}
\title{Tabla Reactable Formateada UES}
\usage{
rt_tabla(
  df,
  fijas = NULL,
  grupos = NULL,
  titulos = NULL,
  filtrar = TRUE,
  barras = NULL,
  color_barra = c("red3", "yellow2", "green3", "blue2", "purple4"),
  destacar_col = NULL,
  color_destacar = "gray",
  cols_porcentaje = NULL,
  destacar_row = NULL,
  highlight_color = "khaki"
)
}
\arguments{
\item{df}{Data frame a mostrar.}

\item{fijas}{Columnas fijas a la izquierda.}

\item{grupos}{Lista nombrada con grupos de columnas.}

\item{titulos}{Vector con nombres personalizados para columnas.}

\item{filtrar}{Si TRUE activa el buscador.}

\item{barras}{Columnas donde aplicar barras.}

\item{color_barra}{Paleta de colores para las barras.}

\item{destacar_col}{Columnas a resaltar.}

\item{color_destacar}{Color de resaltado de columnas.}

\item{cols_porcentaje}{Columnas interpretadas como porcentaje.}

\item{destacar_row}{Filas a resaltar (por valor en la primera columna).}

\item{highlight_color}{Color de fondo para la fila destacada.}
}
\value{
Widget reactable.
}
\description{
Genera tablas avanzadas con \code{reactable} incluyendo columnas fijas, barras,
porcentajes, estilos personalizados y filas destacadas.
}
