% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar_titulos.R
\name{validar_titulos}
\alias{validar_titulos}
\title{Validar y opcionalmente ordenar columnas}
\usage{
validar_titulos(df, titulos, ordenar = FALSE)
}
\arguments{
\item{df}{Data frame a validar.}

\item{titulos}{Vector con nombres de columnas requeridas.}

\item{ordenar}{Logical. Si TRUE, reordena el data frame según \code{titulos}.}
}
\value{
El data frame original o reordenado.
}
\description{
Esta función verifica si un data frame contiene los nombres de columnas
indicados en \code{titulos}. Si \code{ordenar = TRUE}, reordena el data frame en el
mismo orden solicitado.
}
