% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_estadistico.R
\name{mapa_estadistico}
\alias{mapa_estadistico}
\title{Genera un mapa estadístico interactivo para las comunas}
\usage{
mapa_estadistico(
  df_mapa,
  nombre_comuna,
  valor,
  titulo = "Mapa estadístico",
  tipo = c("continuo", "discreto"),
  colores = c("red1", "yellow1", "green3"),
  titulo_leyenda = "Valor",
  comunas_requeridas = c("Chépica", "Chimbarongo", "Codegua", "Coinco", "Coltauco",
    "Doñihue", "Graneros", "La Estrella", "Las Cabras", "Litueche", "Lolol", "Machalí",
    "Malloa", "Marchihue", "Mostazal", "Nancagua", "Navidad", "Olivar", "Palmilla",
    "Paredones", "Peralillo", "Peumo", "Pichidegua", "Pichilemu", "Placilla", "Pumanque",
    "Quinta de Tilcoco", "Rancagua", "Rengo", "Requínoa", "San Fernando", "San Vicente",
    "Santa Cruz"),
  tooltip_text = NULL
)
}
\arguments{
\item{df_mapa}{Data frame con la información a mapear.}

\item{nombre_comuna}{Nombre de la columna con el nombre de la comuna (string).}

\item{valor}{Nombre de la columna con el valor numérico a representar (string).}

\item{titulo}{Título del mapa.}

\item{tipo}{Tipo de escala: "continuo" o "discreto".}

\item{colores}{Vector de colores para la escala, ej: c("red1", "yellow1", "green3")}

\item{titulo_leyenda}{Título de la leyenda.}

\item{comunas_requeridas}{vector opcional para forzar comunas específicas, en este caso ues::comunas_06.}

\item{tooltip_text}{Texto plantilla opcional para personalizar el tooltip, ej: "Comuna: {nombre_comuna}<br>N° de casos: {format(n, big.mark = '.', decimal.mark = ',')}<br>Tasa: {format(round(valor, 2), decimal.mark = ',')}"}
\value{
Un objeto plotly.
}
\description{
Genera un mapa estadístico interactivo para las comunas
}
